Additional fix for the previous commit changing data type for m_lenth in SimpleCString from uint to size_t.

+++ sql/item.h
@@ -156,7 +156,7 @@ public:
 class NameString: public SimpleCString
 {
 private:
-  void set_or_copy(const char *str, uint length, bool is_null_terminated)
+  void set_or_copy(const char *str, size_t length, bool is_null_terminated)
   {
     if (is_null_terminated)
       set(str, length);
